# docs/swagger.yaml
basePath: /
definitions:
...
  dto.AuthTokensResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
  dto.ChangePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        type: string
    type: object
...
  dto.ForgotPasswordRequest:
    properties:
      login:
        type: string
    type: object
  dto.LoginRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
...
  dto.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  dto.RegisterResponse:
    properties:
      registration_id:
        type: string
    type: object
  dto.ResendCodeRequest:
    properties:
      email:
        type: string
      registration_id:
        type: string
    type: object
  dto.ResetPasswordRequest:
    properties:
      code:
        type: string
      email:
        type: string
      new_password:
        type: string
    type: object
...
  dto.VerifyEmailRequest:
    properties:
      code:
        type: string
      email:
        type: string
      registration_id:
        type: string
      username:
        type: string
    type: object
...
  handler.googleLoginRequest:
    properties:
      token:
        type: string
    type: object
  handler.refreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  response.ValidationErrorResponse:
    properties:
      error:
        type: string
      fields:
        additionalProperties:
          type: string
        type: object
    type: object
info:
  contact: {}
  description: API Server for Mobile Application "Meetuper"
  title: Offline Meetup API
  version: "1.0"
paths:
  /v1/auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Принимает login (email или username) и всегда отвечает 202 — независимо
        от того, существует ли такой аккаунт, чтобы эндпоинт не подсказывал, какие
        аккаунты зарегистрированы. Код для сброса отправляется на email, только если
        аккаунт найден.
      parameters:
      - description: Login
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ValidationErrorResponse'
      summary: Запросить сброс пароля
      tags:
      - Auth
  /v1/auth/google:
    post:
      consumes:
      - application/json
      description: Принимает ID Token от Google SDK, возвращает JWT.
      parameters:
      - description: Google ID Token
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.googleLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Вход через Google
      tags:
      - Auth
  /v1/auth/login:
    post:
      consumes:
      - application/json
      description: Принимает login (email или username — определяется по наличию '@')
        и пароль, возвращает пару токенов. Неизвестный login и неверный пароль дают
        одинаковую ошибку.
      parameters:
      - description: Login, пароль
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthTokensResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Вход по email/username и паролю
      tags:
      - Auth
  /v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Отзывает переданный refresh_token. Идемпотентно.
      parameters:
      - description: Refresh token
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.refreshRequest'
      responses:
        "204":
          description: No Content
      summary: Выход
      tags:
      - Auth
  /v1/auth/me:
    get:
      description: Возвращает ID, email, роль и статус.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить данные моего аккаунта
      tags:
      - Auth
  /v1/auth/password:
    patch:
      consumes:
      - application/json
      description: Меняет пароль текущего аккаунта. Если у аккаунта ещё нет пароля
        (вход был только через Google/Telegram), current_password не проверяется.
        Все refresh-токены пользователя отзываются, включая тот, которым выполнен
        этот запрос — клиент должен быть готов заново выполнить вход (или получить
        401 при следующем refresh).
      parameters:
      - description: Текущий и новый пароль
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePasswordRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Сменить пароль
      tags:
      - Auth
  /v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Принимает refresh_token, ротирует его и возвращает новую пару токенов.
      parameters:
      - description: Refresh token
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.refreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthTokensResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Обновить токены
      tags:
      - Auth
  /v1/auth/register:
    post:
      consumes:
      - application/json
      description: Создаёт «ожидающую» регистрацию и отправляет код подтверждения
        на email. Пользователь в БД не создаётся до подтверждения. Ответ одинаков
        независимо от того, существует ли уже аккаунт с таким email. Возвращает registration_id
        — его нужно прислать обратно в /verify-email и /resend-code.
      parameters:
      - description: Email, username, пароль
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/dto.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ValidationErrorResponse'
      summary: Регистрация по email и паролю
      tags:
      - Auth
  /v1/auth/resend-code:
    post:
      consumes:
      - application/json
      description: Генерирует новый код для незавершённой регистрации, адресованной
        парой (email, registration_id). Отвечает 202 и в том случае, когда такой незавершённой
        регистрации нет — чтобы эндпоинт не подсказывал, какие email заняты.
      parameters:
      - description: Email и registration_id из /register
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.ResendCodeRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ValidationErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Отправить код подтверждения повторно
      tags:
      - Auth
  /v1/auth/reset-password:
    post:
      consumes:
      - application/json
      description: Проверяет код из письма и устанавливает новый пароль. Не логинит
        — все refresh-токены пользователя отзываются.
      parameters:
      - description: Email, код, новый пароль
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ValidationErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Завершить сброс пароля
      tags:
      - Auth
...
  /v1/auth/verify-email:
    post:
      consumes:
      - application/json
      description: Проверяет код из письма, создаёт аккаунт (или добавляет пароль
        к существующему) и возвращает пару токенов. Требует registration_id из ответа
        /register. Если username заняли, пока код был в пути, вернётся 409 — повторите
        запрос с другим username.
      parameters:
      - description: Email, registration_id, код, опционально другой username
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthTokensResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ValidationErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Подтвердить email и завершить регистрацию
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: Bearer <your-token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"